+++ src/createSignedTx.ts
@@ -1,9 +1,4 @@
-import {
-  createMetadata,
-  Options,
-  sanitizeOptions,
-  UnsignedTransaction,
-} from './util';
+import { createMetadata, OptionsWithMeta, UnsignedTransaction } from './util';
 
 /**
  * Serialize a signed transaction in a format that can be submitted over the
@@ -17,15 +12,10 @@ import {
 export function createSignedTx(
   unsigned: UnsignedTransaction,
   signature: string,
-  options?: Partial<Options>
+  options: OptionsWithMeta
 ): string {
-  const { metadata, registry } = sanitizeOptions({
-    // FIXME `options` has a metadata field, `unsigned` has a metadata field,
-    // so which one should take precedence? For now, it's `options`.
-    metadata: unsigned.metadataRpc,
-    ...options,
-  });
-  registry.setMetadata(createMetadata(registry, metadata));
+  const { metadataRpc, registry } = options;
+  registry.setMetadata(createMetadata(registry, metadataRpc));
 
   const extrinsic = registry.createType(
     'Extrinsic',