+++ src/decode/decodeUnsignedTx.ts
@@ -4,8 +4,7 @@
 
 import {
   createMetadata,
-  Options,
-  sanitizeOptions,
+  OptionsWithMeta,
   toTxMethod,
   TxInfo,
   UnsignedTransaction,
@@ -21,35 +20,11 @@ export type DecodedUnsignedTx = TxInfo;
  */
 export function decodeUnsignedTx(
   unsigned: UnsignedTransaction,
-  options: Options
-): DecodedUnsignedTx;
-
-/**
- * Parse the transaction information from an unigned transaction offline.
- *
- * @deprecated Prefer passing an `options` object as second argument.
- * @param unsigned - The JSON representing the unsigned transaction.
- * @param metadataRpc - The SCALE-encoded metadata, as a hex string. Can be
- * retrieved via the RPC call `state_getMetadata`.
- * @param ss58Format - The SS-58 address encoding to return.
- */
-export function decodeUnsignedTx(
-  unsigned: UnsignedTransaction,
-  metadataRpc: string,
-  ss58Format: number
-): DecodedUnsignedTx;
-
-export function decodeUnsignedTx(
-  unsigned: UnsignedTransaction,
-  metadataOrOptions: string | Options,
-  _ss58Format?: number
+  options: OptionsWithMeta
 ): DecodedUnsignedTx {
-  const { metadata, registry } = sanitizeOptions(
-    metadataOrOptions,
-    _ss58Format
-  );
+  const { metadataRpc, registry } = options;
 
-  registry.setMetadata(createMetadata(registry, metadata));
+  registry.setMetadata(createMetadata(registry, metadataRpc));
 
   const methodCall = registry.createType('Call', unsigned.method);
   const method = toTxMethod(registry, methodCall);
@@ -62,7 +37,7 @@ export function decodeUnsignedTx(
       .toNumber(),
     eraPeriod: registry.createType('MortalEra', unsigned.era).period.toNumber(),
     genesisHash: unsigned.genesisHash,
-    metadataRpc: metadata,
+    metadataRpc,
     method,
     nonce: registry.createType('Compact<Index>', unsigned.nonce).toNumber(),
     specVersion: registry.createType('u32', unsigned.specVersion).toNumber(),