+++ src/methods/session/setKeys.spec.ts
@@ -1,6 +1,7 @@
 import {
   TEST_BASE_TX_INFO,
   TEST_METHOD_ARGS,
+  TEST_OPTIONS,
   testBaseTxInfo,
 } from '../../util';
 import { setKeys } from './setKeys';
@@ -9,7 +10,8 @@ describe('session::setKeys', () => {
   it('should work', () => {
     const unsigned = setKeys(
       TEST_METHOD_ARGS.session.setKeys,
-      TEST_BASE_TX_INFO
+      TEST_BASE_TX_INFO,
+      TEST_OPTIONS
     );
 
     testBaseTxInfo(unsigned);