+++ src/methods/staking/validate.spec.ts
@@ -1,6 +1,7 @@
 import {
   TEST_BASE_TX_INFO,
   TEST_METHOD_ARGS,
+  TEST_OPTIONS,
   testBaseTxInfo,
 } from '../../util';
 import { validate } from './validate';
@@ -9,7 +10,8 @@ describe('staking::validate', () => {
   it('should work', () => {
     const unsigned = validate(
       TEST_METHOD_ARGS.staking.validate,
-      TEST_BASE_TX_INFO
+      TEST_BASE_TX_INFO,
+      TEST_OPTIONS
     );
 
     testBaseTxInfo(unsigned);