+++ src/methods/vesting/vestOther.spec.ts
@@ -1,6 +1,7 @@
 import {
   TEST_BASE_TX_INFO,
   TEST_METHOD_ARGS,
+  TEST_OPTIONS,
   testBaseTxInfo,
 } from '../../util';
 import { vestOther } from './vestOther';
@@ -9,7 +10,8 @@ describe('vesting::vestOther', () => {
   it('should work', () => {
     const unsigned = vestOther(
       TEST_METHOD_ARGS.vesting.vestOther,
-      TEST_BASE_TX_INFO
+      TEST_BASE_TX_INFO,
+      TEST_OPTIONS
     );
 
     testBaseTxInfo(unsigned);