+++ src/util/index.ts
@@ -1,6 +1,5 @@
 export * from './constants';
 export * from './metadata';
 export * from './method';
-export * from './options';
 export * from './testUtil';
 export * from './types';